+++ dune/grid/io/file/vtk/vtkwriter.hh
@@ -1281,7 +1281,7 @@ namespace Dune
       {
         {
           std::shared_ptr<VTK::DataArrayWriter> p3
-            (writer.makeArrayWriter("types", 1, ncells, VTK::Precision::uint32));
+            (writer.makeArrayWriter("types", 1, ncells, VTK::Precision::uint8));
 
           if(!p3->writeIsNoop())
           {
